Cleaning to obtain fstat v2.0 version
