--- try_bayes/clean.x
@@ -1,1 +1,0 @@
-rm *.pdf *.txt OUTPUT
